+++ sdk/partner/PartnerClient.js
@@ -10,6 +10,7 @@ const Webhook = require("./Webhook/WebhookPartnerClient");
 
 const { FDKClientValidationError } = require("../common/FDKError");
 const { execute } = require("./PartnerAPIClient");
+const PartnerConfig = require("./PartnerConfig");
 
 /**
  * Represents the client for the partner APIs.
@@ -23,13 +24,18 @@ class PartnerClient {
    * @param {import("./PartnerConfig")} config - The configuration for the
    *   partner client.
    */
-  constructor(config) {
-    this.config = config;
-    this.fileStorage = new FileStorage(config);
-    this.lead = new Lead(config);
-    this.logistics = new Logistics(config);
-    this.theme = new Theme(config);
-    this.webhook = new Webhook(config);
+  constructor(config, options) {
+    if (config instanceof PartnerConfig) {
+      this.config = config;
+    } else {
+      let partnerConfig = new PartnerConfig(config, options);
+      this.config = partnerConfig;
+    }
+    this.fileStorage = new FileStorage(this.config);
+    this.lead = new Lead(this.config);
+    this.logistics = new Logistics(this.config);
+    this.theme = new Theme(this.config);
+    this.webhook = new Webhook(this.config);
   }
 
   /**
@@ -57,6 +63,14 @@ class PartnerClient {
     return await execute(this.config, method, url, query, body, headers, {
       responseHeaders,
     });
+  }
+
+  getAccesstokenObj(options) {
+    return this.config.oauthClient.getAccesstokenObj(options);
+  }
+
+  setToken(token) {
+    this.config.oauthClient.setToken(token);
   }
 }