+++ sdk/platform/Payment/PaymentPlatformApplicationValidator.d.ts
@@ -114,9 +114,8 @@ export = PaymentPlatformApplicationValidator;
  */
 /**
  * @typedef GetPosPaymentModeRoutesParam
- * @property {PaymentPlatformModel.OrderingSource} [xOrderingSource] - Optional
- *   header to identify the ordering source used to determine\ applicable
- *   payment options for business unit.
+ * @property {string} [xOrderingSource] - Optional header to identify the
+ *   ordering source used to determine\ applicable payment options for business unit.
  * @property {number} amount - Payable amount.
  * @property {string} [cartId] - Identifier of the cart.
  * @property {string} pincode - The PIN Code of the destination address, e.g. 400059
@@ -496,11 +495,10 @@ type GetPaymentSessionParam = {
 };
 type GetPosPaymentModeRoutesParam = {
     /**
-     * - Optional
-     * header to identify the ordering source used to determine\ applicable
-     * payment options for business unit.
+     * - Optional header to identify the
+     * ordering source used to determine\ applicable payment options for business unit.
      */
-    xOrderingSource?: PaymentPlatformModel.OrderingSource;
+    xOrderingSource?: string;
     /**
      * - Payable amount.
      */